Commit for backup: partial implementation of new plant structure
I'm leaving for a couple of days and I want to push this incomplete
work to GitHub.

See `feedbax.plant.MuscledArm` and `feedbax.mechanics.Mechanics` for
most of the relevant changes. Mainly, attempting to separate the
direct state updates from the ODE updates, and set up the ODE
updates with `dynamics_spec` in a similar way to `model_spec`.